fix: asLong, asULong

+++ src/tinyredis_util/package.d
@@ -167,12 +167,12 @@ unittest {
    assert(!redis.asBool!("delete_me"));
 }
 
-int asLong(string K)(Redis redis) {
+long asLong(string K)(Redis redis) {
    string reply = redis.send!string("GET", K);
    return conv!long(reply);
 }
 
-int asULong(string K)(Redis redis) {
+ulong asULong(string K)(Redis redis) {
    string reply = redis.send!string("GET", K);
    return conv!ulong(reply);
 }
